% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_name_usage.R
\name{cb_name_usage}
\alias{cb_name_usage}
\title{Title}
\usage{
cb_name_usage(
  q = NULL,
  key = "304862",
  id = NULL,
  name = NULL,
  scientificName = NULL,
  authorship = NULL,
  code = NULL,
  rank = NULL,
  status = NULL,
  verbose = NULL,
  superkingdom = NULL,
  kingdom = NULL,
  subkingdom = NULL,
  superphylum = NULL,
  phylum = NULL,
  subphylum = NULL,
  superclass = NULL,
  class = NULL,
  subclass = NULL,
  superorder = NULL,
  order = NULL,
  suborder = NULL,
  superfamily = NULL,
  family = NULL,
  subfamily = NULL,
  tribe = NULL,
  subtribe = NULL,
  genus = NULL,
  subgenus = NULL,
  section = NULL,
  species = NULL,
  user = NULL,
  pwd = NULL,
  curlopts = list()
)
}
\arguments{
\item{q}{}

\item{key}{}

\item{id}{}

\item{name}{}

\item{scientificName}{}

\item{authorship}{}

\item{code}{}

\item{rank}{}

\item{status}{}

\item{verbose}{}

\item{superkingdom}{}

\item{kingdom}{}

\item{subkingdom}{}

\item{superphylum}{}

\item{phylum}{}

\item{subphylum}{}

\item{superclass}{}

\item{class}{}

\item{subclass}{}

\item{superorder}{}

\item{order}{}

\item{suborder}{}

\item{superfamily}{}

\item{family}{}

\item{subfamily}{}

\item{tribe}{}

\item{subtribe}{}

\item{genus}{}

\item{subgenus}{}

\item{section}{}

\item{species}{}

\item{user}{}

\item{pwd}{}

\item{curlopts}{}
}
\description{
Title
}
