% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.r
\name{commas_to_periods}
\alias{commas_to_periods}
\title{Convert commas to periods in lat/long data}
\usage{
commas_to_periods(dataframe)
}
\arguments{
\item{dataframe}{A data.frame}
}
\description{
Convert commas to periods in lat/long data
}
\keyword{internal}

