% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_gist}
\alias{create_gist}
\title{Function that takes a list of files and creates payload for API}
\usage{
create_gist(...)
}
\description{
This function is defunct.  See the package gistr for similar functionality.
}
\keyword{internal}

