% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_list_funs.R
\name{dataset_list_funs}
\alias{dataset_list_funs}
\alias{dataset_duplicate}
\alias{dataset_noendpoint}
\title{List datasets that are deleted or have no endpoint.}
\usage{
dataset_duplicate(limit = 20, start = NULL, curlopts = list(http_version = 2))

dataset_noendpoint(limit = 20, start = NULL, curlopts = list(http_version = 2))
}
\arguments{
\item{limit}{Controls the number of results in the page.}

\item{start}{Determines the start for the search results.}

\item{curlopts}{options passed on to \link[crul:HttpClient]{crul::HttpClient}.}
}
\value{
A \code{list}.
}
\description{
List datasets that are deleted or have no endpoint.
}
\details{
Get a list of deleted datasets or datasets with no endpoint. You get the full
and no parameters aside from \code{limit} and \code{start} are accepted.
}
\examples{
\dontrun{
dataset_noendpoint(limit=3)
}
}
