% Generated by roxygen2 (4.0.1): do not edit by hand
\name{datasetparser}
\alias{datasetparser}
\title{Parser for gbif data}
\usage{
datasetparser(input, minimal = TRUE)
}
\arguments{
\item{input}{A list}

\item{minimal}{Get minimal input, default to TRUE}
}
\description{
Parser for gbif data
}
\keyword{internal}

