\name{density_spplist}
\alias{density_spplist}
\title{The density web service provides access to records showing the density
   of occurrence records from the GBIF Network by one-degree cell.}
\usage{
  density_spplist()
}
\description{
  Note: TThis function is deprecated. There is no longer a
  similar function.
}

