\name{densitylist}
\alias{densitylist}
\title{The density web service provides access to records showing the density
   of occurrence records from the GBIF Network by one-degree cell.}
\usage{
  densitylist()
}
\description{
  Note: This function is deprecated. There is no longer a
  similar functio
}

