% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_user_dl.R
\name{dl_exists}
\alias{dl_exists}
\title{does a download exist already on the GBIF's user account remotely}
\usage{
dl_exists(pred, preds)
}
\arguments{
\item{pred}{a download request predicate, the output from
\code{\link[=occ_download_prep]{occ_download_prep()}}}

\item{preds}{output from \code{\link[=dl_predicates]{dl_predicates()}}}
}
\description{
does a download exist already on the GBIF's user account remotely
}
\examples{
dprep <- occ_download_prep("catalogNumber = Bird.27847588", "year = 1978",
  "month = 5")
preds <- dl_predicates(dl_user())
dl_exists(pred = dprep, preds)
}
\keyword{internal}
