% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_user_dl.R
\name{dl_match}
\alias{dl_match}
\title{Match user supplied predicate to their GBIF downloads}
\usage{
dl_match(pred, preds, age = 30)
}
\arguments{
\item{pred}{() a predicate, object of class \code{occ_download_prep}, output
from \code{\link[=occ_download_prep]{occ_download_prep()}}}

\item{preds}{(data.frame) output from \code{\link[=dl_predicates]{dl_predicates()}}}

\item{age}{(integer) number of days after which you want a new
download. default: 30}
}
\value{
if no matches list, with \code{match=FALSE}. if any matches returns
list with \code{match=TRUE}, and key, created, and totalRecords fields. If a
match was found but it was created more than
}
\description{
Match user supplied predicate to their GBIF downloads
}
\note{
\itemize{
\item remove any with job status: "PREPARING", "KILLED", "CANCELLED",
"FILE_ERASED" (leaving only "SUCCEEDED")
\item check for any matches. if any matches, then:
\itemize{
\item sort by most recent request
\item if the query age created is less than the \code{age} parameter, return \code{TRUE}
and the download key
}
}
}
\examples{
\dontrun{
preds <- dl_predicates(dl_user())
preds

# not matched
dprep <- occ_download_prep(pred("catalogNumber", "Bird.27847588"),
  pred("year", 1978))
dl_match(pred = dprep, preds)

# match but expired
dprep <- occ_download_prep(pred("catalogNumber", "Bird.27847588"))
dl_match(pred = dprep, preds)

# match, and not expired
dprep <- occ_download_prep(pred_within("POLYGON((-14 42, 9 38, -7 26, -14 42))"),
  pred_gte("elevation", 5000))
dl_match(pred = dprep, preds, age = 90)
}
}
\keyword{internal}
