% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_user_dl.R
\name{dl_user}
\alias{dl_user}
\title{Fetch all downloads for a user}
\usage{
dl_user(user = NULL, pwd = NULL, curlopts = list())
}
\arguments{
\item{user}{(character) User name within GBIF's website. Required. See
"Authentication" below}

\item{pwd}{(character) User password within GBIF's website. Required. See
"Authentication" below}

\item{curlopts}{list of named curl options passed on to
\code{\link[crul]{HttpClient}}. see \code{\link[curl]{curl_options}}
for curl options}
}
\description{
Fetch all downloads for a user
}
\examples{
\dontrun{
x <- dl_user()
x
}
}
\keyword{internal}
