% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_predicate_dsl.R
\name{download_predicate_dsl}
\alias{download_predicate_dsl}
\alias{pred}
\alias{pred_gt}
\alias{pred_gte}
\alias{pred_lt}
\alias{pred_lte}
\alias{pred_not}
\alias{pred_like}
\alias{pred_within}
\alias{pred_notnull}
\alias{pred_or}
\alias{pred_and}
\alias{pred_in}
\title{Download predicate DSL}
\usage{
pred(key, value)

pred_gt(key, value)

pred_gte(key, value)

pred_lt(key, value)

pred_lte(key, value)

pred_not(key, value)

pred_like(key, value)

pred_within(value)

pred_notnull(key)

pred_or(..., .list = list())

pred_and(..., .list = list())

pred_in(key, value)
}
\arguments{
\item{key}{(character) the key for the predicate (used in \code{pred()},
\code{pred_multi()}, and \code{preds()}). See "Keys" below}

\item{value}{(various) the value for the predicate (used in \code{pred()},
\code{pred_multi()}, and \code{preds()})}

\item{..., .list}{For \code{pred_or()} or \code{pred_and()}, one or more objects of class
\code{occ_predicate}, created by \code{pred()}, etc.}
}
\description{
Download predicate DSL
}
\section{predicate methods and their equivalent types}{

\itemize{
\item equals (=): \code{pred}
\item and (&): \code{pred_and}
\item or (|): \code{pred_or}
\item lessThan (<): \code{pred_lt}
\item lessThanOrEquals (<=): \code{pred_lte}
\item greaterThan (>): \code{pred_gt}
\item greaterThanOrEquals (>=): \code{pred_gte}
\item in: \code{pred_in}
\item within: \code{pred_within} (only geospatial, accepts only a WKT string)
\item not (!): \code{pred_not}
\item like: \code{pred_like}
\item isNotNull: \code{pred_notnull}
}
}

\section{What happens internally}{

Internally, the input to \verb{pred*} functions turns into JSON to be sent to
GBIF. For example, \code{pred_in("taxonKey", c(2480946, 5229208))} will give:\preformatted{'\{
   "type": "or",
   "predicates": [
     \{
      "type": "equals",
      "key": "TAXON_KEY",
      "value": "2480946"
     \},
     \{
      "type": "equals",
      "key": "TAXON_KEY",
      "value": "5229208"
     \}
   ]
\}'
}
}

\section{Keys}{


Acceptable arguments to the \code{key} parameter are (with the version of
the key in parens that must be sent if you pass the query via the \code{body}
parameter; see below for examples):
\itemize{
\item taxonKey (TAXON_KEY)
\item scientificName (SCIENTIFIC_NAME)
\item country (COUNTRY)
\item publishingCountry (PUBLISHING_COUNTRY)
\item hasCoordinate (HAS_COORDINATE)
\item hasGeospatialIssue (HAS_GEOSPATIAL_ISSUE)
\item typeStatus (TYPE_STATUS)
\item recordNumber (RECORD_NUMBER)
\item lastInterpreted (LAST_INTERPRETED)
\item continent (CONTINENT)
\item geometry (GEOMETRY)
\item basisOfRecord (BASIS_OF_RECORD)
\item datasetKey (DATASET_KEY)
\item eventDate (EVENT_DATE)
\item catalogNumber (CATALOG_NUMBER)
\item year (YEAR)
\item month (MONTH)
\item decimalLatitude (DECIMAL_LATITUDE)
\item decimalLongitude (DECIMAL_LONGITUDE)
\item elevation (ELEVATION)
\item depth (DEPTH)
\item institutionCode (INSTITUTION_CODE)
\item collectionCode (COLLECTION_CODE)
\item issue (ISSUE)
\item mediatype (MEDIA_TYPE)
\item recordedBy (RECORDED_BY)
}
}

\examples{
pred("taxonKey", 3119195)
pred_gt("elevation", 5000)
pred_gte("elevation", 5000)
pred_lt("elevation", 1000)
pred_lte("elevation", 1000)
pred_within("POLYGON((-14 42, 9 38, -7 26, -14 42))")
pred_or(pred_lte("year", 1989), pred("year", 2000))
pred_and(pred_lte("year", 1989), pred("year", 2000))
pred_in("taxonKey", c(2977832, 2977901, 2977966, 2977835))
pred_in("basisOfRecord", c("MACHINE_OBSERVATION", "HUMAN_OBSERVATION"))
pred_not("catalogNumber", "cat1")
pred_like("catalogNumber", "PAPS5-560\%")
pred_notnull("issue")
pred("basisOfRecord", "LITERATURE")
pred("hasCoordinate", TRUE)
pred("hasGeospatialIssue", FALSE)
pred_within("POLYGON((-14 42, 9 38, -7 26, -14 42))")
pred_or(pred("taxonKey", 2977832), pred("taxonKey", 2977901),
  pred("taxonKey", 2977966))
pred_in("taxonKey", c(2977832, 2977901, 2977966, 2977835))
}
