% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_user_dl.R
\name{DownloadMatch}
\alias{DownloadMatch}
\title{DownloadMatch}
\description{
download match handler
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pred}}{a predicate}

\item{\code{preds}}{(data.frame) preds}

\item{\code{age}}{(integer) age}

\item{\code{output}}{(list) the output}

\item{\code{most_recent}}{(character) most recent matching date}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{message}}{\itemize{
\item make an output message
}}

\item{\code{matched}}{\itemize{
\item match boolean
}}

\item{\code{expired}}{\itemize{
\item expired boolean
}}

\item{\code{key}}{\itemize{
\item download key
}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{DownloadMatch$print()}}
\item \href{#method-new}{\code{DownloadMatch$new()}}
\item \href{#method-check_matches}{\code{DownloadMatch$check_matches()}}
\item \href{#method-pred_match}{\code{DownloadMatch$pred_match()}}
\item \href{#method-clone}{\code{DownloadMatch$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print method for the \code{DownloadMatch} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DownloadMatch$print(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{self}

\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{DownloadMatch} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DownloadMatch$new(pred, preds, age = 30)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pred}}{() a predicate, object of class \code{occ_download_prep}, output
from \code{\link[=occ_download_prep]{occ_download_prep()}}}

\item{\code{preds}}{(data.frame) output from \code{\link[=dl_predicates]{dl_predicates()}}}

\item{\code{age}}{(integer) number of days after which you want a new
download. default: 30}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{DownloadMatch} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_matches"></a>}}
\if{latex}{\out{\hypertarget{method-check_matches}{}}}
\subsection{Method \code{check_matches()}}{
Check for matches
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DownloadMatch$check_matches()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing returned; adds job (\code{x}) to the queue
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pred_match"></a>}}
\if{latex}{\out{\hypertarget{method-pred_match}{}}}
\subsection{Method \code{pred_match()}}{
create output list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DownloadMatch$pred_match(
  match = FALSE,
  key = NULL,
  created = NULL,
  n = NULL,
  expired = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{match}}{(logical) a boolean, \code{TRUE} if a match found}

\item{\code{key}}{(character) the download key, \code{NA_character_} if no
match found}

\item{\code{created}}{(character) date created}

\item{\code{n}}{(integer) number of records found}

\item{\code{expired}}{(logical) a boolean, if older than the age given}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DownloadMatch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
