% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_queue.R
\docType{data}
\name{DownReq}
\alias{DownReq}
\title{Download request}
\description{
Download request
}
\details{
\strong{Methods}
\describe{
\item{\code{new(...)}}{
initialize request
}
\item{\code{run(...)}}{
execute http request
}
\item{\code{status(...)}}{
check http request status
}
}
}
\examples{
\dontrun{
res <- DownReq$new(occ_download('taxonKey = 3119195', "year = 1991"))
res
res$req
res$run()
# (requests <- GbifQueue$new())
# res$run(keep_track = TRUE)
# requests
res$status()
}
}
\keyword{internal}
