% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.r
\name{gbif_capwords}
\alias{gbif_capwords}
\title{Capitalize the first letter of a character string.}
\usage{
gbif_capwords(s, strict = FALSE, onlyfirst = FALSE)
}
\arguments{
\item{s}{A character string}

\item{strict}{Should the algorithm be strict about capitalizing. Defaults to FALSE.}

\item{onlyfirst}{Capitalize only first word, lowercase all others. Useful for
taxonomic names.}
}
\description{
Capitalize the first letter of a character string.
}
\examples{
gbif_capwords(c("using AIC for model selection"))
gbif_capwords(c("using AIC for model selection"), strict=TRUE)
}
\keyword{internal}

