% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_issues.R
\name{gbif_issues}
\alias{gbif_issues}
\title{List all GBIF issues and their codes.}
\source{
https://gbif.github.io/gbif-api/apidocs/org/gbif/api/vocabulary/OccurrenceIssue.html
https://gbif.github.io/gbif-api/apidocs/org/gbif/api/vocabulary/NameUsageIssue.html
}
\usage{
gbif_issues()
}
\description{
Returns a data.frame of all GBIF issues with the following columns:
\itemize{
\item \code{code}: issue short code, e.g. \code{gass84}
\item \code{code}: issue full name, e.g. \code{GEODETIC_DATUM_ASSUMED_WGS84}
\item \code{description}: issue description
\item \code{type}: issue type, either related to \code{occurrence} or \code{name}
}
}
