% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_issues_lookup.r
\name{gbif_issues_lookup}
\alias{gbif_issues_lookup}
\title{Lookup issue definitions and short codes}
\usage{
gbif_issues_lookup(issue = NULL, code = NULL)
}
\arguments{
\item{issue}{Full name of issue, e.g, CONTINENT_COUNTRY_MISMATCH}

\item{code}{An issue short code, e.g. 'ccm'}
}
\description{
Lookup issue definitions and short codes
}
\examples{
gbif_issues_lookup(issue = 'CONTINENT_COUNTRY_MISMATCH')
gbif_issues_lookup(code = 'ccm')
gbif_issues_lookup(issue = 'COORDINATE_INVALID')
gbif_issues_lookup(code = 'cdiv')
}
