\name{gbif_lookup}
\alias{gbif_lookup}
\title{Lookup names in the GBIF backbone taxonomy.}
\usage{
  gbif_lookup(name, kingdom = NULL)
}
\arguments{
  \item{name}{}

  \item{kingdom}{}
}
\description{
  Lookup names in the GBIF backbone taxonomy.
}
\examples{
\dontrun{
gbif_lookup(name='Helianthus annuus', kingdom='plants')
}
}

