% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gbif_names}
\alias{gbif_names}
\title{View highlighted terms in name results from GBIF.}
\usage{
gbif_names(input, output = NULL, browse = TRUE)
}
\arguments{
\item{input}{Input output from occ_search}

\item{output}{Output folder path. If not given uses temporary folder.}

\item{browse}{(logical) Browse output (default: TRUE)}
}
\description{
View highlighted terms in name results from GBIF.
}
\examples{
\donttest{
(out <- name_lookup(query='canada', hl=TRUE, limit=5))
gbif_names(out)
gbif_names(name_lookup(query='snake', hl=TRUE))
gbif_names(name_lookup(query='bird', hl=TRUE))

# or not highlight
gbif_names(name_lookup(query='bird', limit=200))
}
}

