% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gbif_photos}
\alias{gbif_photos}
\title{View photos}
\usage{
gbif_photos(input = NULL, output = NULL, browse = TRUE)
}
\arguments{
\item{input}{Input output from occ_search}

\item{output}{Output file name and path}

\item{browse}{(logical) Browse output}
}
\description{
View photos
}
\examples{
\dontrun{
(res <- occ_search(mediatype = 'StillImage', return = "media"))
gbif_photos(res)

res <- occ_search(scientificName = "Helianthus", mediatype = 'StillImage', return = "media")
gbif_photos(res)

res <- occ_search(scientificName = "Aves", mediatype = 'StillImage', return = "media", limit=150)
gbif_photos(res)
}
}

