\name{gbifdata.gbifdens}
\alias{gbifdata.gbifdens}
\title{Gbifdens method}
\usage{
  \method{gbifdata}{gbifdens} (input, ...)
}
\arguments{
  \item{input}{Input object from a call to occurrencelist,
  occurrencelist_many, or densitylist.}

  \item{...}{further arguments}
}
\description{
  Gbifdens method
}

