\name{gbifdata.gbiflist_na}
\alias{gbifdata.gbiflist_na}
\title{Gbiflist NA method for when there's no data returned by occurrencelist fxn}
\usage{
  \method{gbifdata}{gbiflist_na} (input)
}
\arguments{
  \item{input}{Input object from a call to occurrencelist,
  occurrencelist_many, or densitylist.}
}
\description{
  Gbiflist NA method for when there's no data returned by
  occurrencelist fxn
}
\seealso{
  gbifdata.gbiflist
}

