\name{gbifdatause}
\alias{gbifdatause}
\title{GBIF datause descriptiom.}
\usage{
  gbifdatause(url = "http://data.gbif.org/tutorial/datasharingagreement")

  gbifdatause(url = "http://data.gbif.org/tutorial/datasharingagreement")
}
\arguments{
  \item{url}{Base url for call, don't change.}

  \item{url}{Base url for call, don't change.}
}
\description{
  GBIF datause descriptiom.

  GBIF datause descriptiom.
}
\examples{
\dontrun{
gbifdatause()
}
\dontrun{
gbifdatause()
}
}

