% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{gbifissues}
\alias{gbifissues}
\title{Table of GBIF issues, with codes used in data output, full issue name, and descriptions.}
\description{
Table has the following fields:
}
\details{
\itemize{
  \item code. Code for issue, making viewing data easier.
  \item issue. Full name of the issue.
  \item description. Description of the issue.
}
}
\keyword{data}

