% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{gbifmap}
\alias{gbifmap}
\title{Get Github credentials from use in console}
\usage{
gbifmap(...)
}
\description{
This function is defunct.  See the package gistr for similar functionality.
}
\keyword{internal}
