% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{gbifmap_dens}
\alias{gbifmap_dens}
\title{Make a simple map to visualize GBIF data density data}
\usage{
gbifmap_dens(...)
}
\description{
This function is defunct.
}
\seealso{
gbifmap
}
\keyword{internal}
