% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gbifxmlToDataFrame}
\alias{gbifxmlToDataFrame}
\title{Code based on the `gbifxmlToDataFrame` function from dismo package
(http://cran.r-project.org/web/packages/dismo/index.html),
by Robert Hijmans, 2012-05-31, License: GPL v3}
\usage{
gbifxmlToDataFrame(doc, format)
}
\arguments{
\item{doc}{A parsed XML document.}

\item{format}{Format to use.}
}
\description{
Code based on the `gbifxmlToDataFrame` function from dismo package
(http://cran.r-project.org/web/packages/dismo/index.html),
by Robert Hijmans, 2012-05-31, License: GPL v3
}
\keyword{internal}

