% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HttpStore.R
\docType{data}
\name{HttpStore}
\alias{HttpStore}
\title{HttpStore}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HttpStore
}
\description{
HttpStore
}
\details{
The original idea was to use dates in returned http response
headers from GBIF, but apparently it seems if they serve a cached
response that the date header is from whenever that response was cached,
so it's not the time/date the response itself is served. Thus, we just
use a timestamp created immediately after a response is returned.
}
\section{methods}{

\itemize{
\item count: number of requests recorded
\item put: record a request
\item pop: delete the last request
\item wait: decide if should sleep or not. sleeps for amount
of time \code{sleep} (seconds)
}
}

\section{wait method parameters}{

\itemize{
\item sleep (integer): seconds to sleep, default: 1 sec
\item report (logical): whether to report sleep time or not as a message,
default: \code{FALSE}
}
}

\examples{
x <- HttpStore$new()
x$count()
x$requests
x$put("Thu, 13 Sep 2018 02:51:08 GMT")
x$requests
x$count()
x$wait()

x$put(format(Sys.time(), format = "\%d \%B \%Y \%H:\%M:\%S", tz="UTC"))
x$wait()

# failure behavior
x <- rgbif:::HttpStore$new()
x$put(45)
x$wait()
x$put('asdfadfafasdfdffsdf')
x$requests
x$wait()
}
\keyword{internal}
