% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbif-package.r
\docType{data}
\name{isocodes}
\alias{isocodes}
\title{Table of country two character ISO codes, and GBIF names}
\description{
\itemize{
\item code. Two character ISO country code.
\item name. Name of country.
\item gbif_name. Name of country used by GBIF - this is the name
you want to use when searching by country in this package.
}
}
\keyword{data}
