% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_issues_lookup.r
\name{name_issues_lookup}
\alias{name_issues_lookup}
\title{Lookup name issue definitions and short codes.}
\usage{
name_issues_lookup(issue = NULL, code = NULL)
}
\arguments{
\item{issue}{Full name of name issue, e.g, CONTINENT_COUNTRY_MISMATCH}

\item{code}{an name issue short code, e.g. ccm}
}
\description{
Lookup name issue definitions and short codes.
}
\examples{
name_issues_lookup(issue = 'BACKBONE_MATCH_NONE')
name_issues_lookup(issue = 'RANK_INVALID')
name_issues_lookup(code = 'bbmn')
name_issues_lookup(code = 'rankinv')
}
