% Generated by roxygen2 (4.0.0): do not edit by hand
\name{namelkupparser}
\alias{namelkupparser}
\title{Parser for name_usage endpoints, for fxns name_lookup and name_backbone}
\usage{
namelkupparser(x)
}
\arguments{
\item{x}{A list.}
}
\description{
Parser for name_usage endpoints, for fxns name_lookup and name_backbone
}
\keyword{internal}

