% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_stats.R
\name{occ_download_stats}
\alias{occ_download_stats}
\title{Occurrence download statistics}
\usage{
occ_download_stats(
  from = NULL,
  to = NULL,
  publishingCountry = NULL,
  datasetKey = NULL,
  publishingOrgKey = NULL,
  limit = NULL,
  offset = NULL,
  curlopts = list(http_version = 2)
)
}
\arguments{
\item{from}{(character) Start date in format \code{YYYY-MM}. Optional.}

\item{to}{(character) End date in format \code{YYYY-MM}. Optional.}

\item{publishingCountry}{(character) ISO 2-letter country code. Optional.}

\item{datasetKey}{(character) Dataset UUID. Optional.}

\item{publishingOrgKey}{(character) Publishing organization UUID. Optional.}

\item{limit}{(integer) Number of results to return. Optional.}

\item{offset}{(integer) Offset for pagination. Optional.}

\item{curlopts}{list of named curl options passed on to \link[crul:HttpClient]{crul::HttpClient}.
See \link[curl:curl_options]{curl::curl_options} for curl options}
}
\value{
A list with two slots:
\itemize{
\item meta: a single row data.frame with columns: \code{offset}, \code{limit},
\code{endofrecords}, \code{count}
\item results: a tibble with the results containing columns: \code{datasetKey},
\code{totalRecords}, \code{numberDownloads}, \code{year}, \code{month}
}
}
\description{
Retrieve statistics about GBIF occurrence downloads. Filters for downloads
matching the provided criteria, then provides counts by year, month and
dataset of the total number of downloads, and the total number of records
included in those downloads.
}
\note{
see \link{downloads} for an overview of GBIF downloads methods
}
\examples{
\dontrun{
# Get summarized download statistics
occ_download_stats()

# Filter by date range and country
occ_download_stats(from = "2023-01", to = "2023-12", publishingCountry = "US")

# Filter by publishing organization iNaturalist
occ_download_stats(publishingOrgKey = "28eb1a3f-1c15-4a95-931a-4af90ecb574d")
}
}
\seealso{
Other downloads: 
\code{\link{download_predicate_dsl}},
\code{\link{occ_download_cached}()},
\code{\link{occ_download_cancel}()},
\code{\link{occ_download_dataset_activity}()},
\code{\link{occ_download_datasets}()},
\code{\link{occ_download_get}()},
\code{\link{occ_download_import}()},
\code{\link{occ_download_list}()},
\code{\link{occ_download_meta}()},
\code{\link{occ_download_queue}()},
\code{\link{occ_download_stats_dataset_records}()},
\code{\link{occ_download_stats_dataset}()},
\code{\link{occ_download_stats_export}()},
\code{\link{occ_download_stats_source}()},
\code{\link{occ_download_stats_user_country}()},
\code{\link{occ_download_wait}()},
\code{\link{occ_download}()}
}
\concept{downloads}
