% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_stats.R
\name{occ_download_stats_dataset_records}
\alias{occ_download_stats_dataset_records}
\title{Downloaded records by dataset}
\usage{
occ_download_stats_dataset_records(
  from = NULL,
  to = NULL,
  publishingCountry = NULL,
  datasetKey = NULL,
  publishingOrgKey = NULL,
  curlopts = list(http_version = 2)
)
}
\arguments{
\item{from}{(character) Start date in format \code{YYYY-MM}. Optional.}

\item{to}{(character) End date in format \code{YYYY-MM}. Optional.}

\item{publishingCountry}{(character) ISO 2-letter country code. Optional.}

\item{datasetKey}{(character) Dataset UUID. Optional.}

\item{publishingOrgKey}{(character) Publishing organization UUID. Optional.}

\item{curlopts}{list of named curl options passed on to \link[crul:HttpClient]{crul::HttpClient}.
See \link[curl:curl_options]{curl::curl_options} for curl options}
}
\value{
A tibble with columns: \code{year}, \code{month}, \code{number_records}. Each row
represents the total number of records downloaded for a given year and month.
}
\description{
Summarize downloaded records by dataset.
}
\note{
see \link{downloads} for an overview of GBIF downloads methods
}
\examples{
\dontrun{
# Get downloaded records by dataset
occ_download_stats_dataset_records()

# Filter by date range
occ_download_stats_dataset_records(from = "2023-01", to = "2023-12")

# Filter by specific dataset
occ_download_stats_dataset_records(
  datasetKey = "50c9509d-22c7-4a22-a47d-8c48425ef4a7"
)
}
}
\seealso{
Other downloads: 
\code{\link{download_predicate_dsl}},
\code{\link{occ_download_cached}()},
\code{\link{occ_download_cancel}()},
\code{\link{occ_download_dataset_activity}()},
\code{\link{occ_download_datasets}()},
\code{\link{occ_download_get}()},
\code{\link{occ_download_import}()},
\code{\link{occ_download_list}()},
\code{\link{occ_download_meta}()},
\code{\link{occ_download_queue}()},
\code{\link{occ_download_stats_dataset}()},
\code{\link{occ_download_stats_export}()},
\code{\link{occ_download_stats_source}()},
\code{\link{occ_download_stats_user_country}()},
\code{\link{occ_download_stats}()},
\code{\link{occ_download_wait}()},
\code{\link{occ_download}()}
}
\concept{downloads}
