% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_stats.R
\name{occ_download_stats_export}
\alias{occ_download_stats_export}
\title{Export summary of occurrence downloads}
\usage{
occ_download_stats_export(
  from,
  to,
  publishingCountry,
  datasetKey = NULL,
  publishingOrgKey = NULL,
  limit = NULL,
  offset = NULL
)
}
\arguments{
\item{from}{(character) Start date in format \code{YYYY-MM}. Required.}

\item{to}{(character) End date in format \code{YYYY-MM}. Required.}

\item{publishingCountry}{(character) ISO 2-letter country code. Required.}

\item{datasetKey}{(character) Dataset UUID. Optional.}

\item{publishingOrgKey}{(character) Publishing organization UUID. Optional.}

\item{limit}{(integer) Number of results to return. Optional.}

\item{offset}{(integer) Offset for pagination. Optional.}
}
\value{
A tibble with download summary data
}
\description{
Export a summary of occurrence downloads. Note that \code{from}, \code{to}, and
\code{publishingCountry} are required parameters for this endpoint.
}
\note{
see \link{downloads} for an overview of GBIF downloads methods
}
\examples{
\dontrun{
# Export download summary (from, to, and publishingCountry are required)
occ_download_stats_export(from = "2023-01", to = "2023-12", publishingCountry = "US")
}
}
\seealso{
Other downloads: 
\code{\link{download_predicate_dsl}},
\code{\link{occ_download_cached}()},
\code{\link{occ_download_cancel}()},
\code{\link{occ_download_dataset_activity}()},
\code{\link{occ_download_datasets}()},
\code{\link{occ_download_get}()},
\code{\link{occ_download_import}()},
\code{\link{occ_download_list}()},
\code{\link{occ_download_meta}()},
\code{\link{occ_download_queue}()},
\code{\link{occ_download_stats_dataset_records}()},
\code{\link{occ_download_stats_dataset}()},
\code{\link{occ_download_stats_source}()},
\code{\link{occ_download_stats_user_country}()},
\code{\link{occ_download_stats}()},
\code{\link{occ_download_wait}()},
\code{\link{occ_download}()}
}
\concept{downloads}
