% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_spellcheck.R
\name{occ_spellcheck}
\alias{occ_spellcheck}
\title{Spell check search term for occurrence searches}
\usage{
occ_spellcheck(...)
}
\arguments{
\item{...}{ignored}
}
\description{
Spell check search term for occurrence searches
}
\keyword{internal}
