% Generated by roxygen2 (4.0.0): do not edit by hand
\name{occurrencedensity}
\alias{occurrencedensity}
\title{Returns summary counts of occurrence records by one-degree cell for a single
  	taxon, country, dataset, data publisher or data network.}
\usage{
occurrencedensity()
}
\description{
This function is deprecated.
}
\seealso{
densitylist
}

