% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{occurrencedensity}
\alias{occurrencedensity}
\title{Returns summary counts of occurrence records by one-degree cell for a single
  	taxon, country, dataset, data publisher or data network.}
\usage{
occurrencedensity()
}
\description{
This function is defunct.
}
\keyword{internal}

