% Generated by roxygen2 (4.0.2): do not edit by hand
\name{occurrencelist_many}
\alias{occurrencelist_many}
\title{occurrencelist_many is the same as occurrencelist, but takes in a vector
of species names.}
\usage{
occurrencelist_many(...)
}
\description{
This function is defunct.
}
\seealso{
occ_search
}
\keyword{internal}

