% Generated by roxygen2 (4.0.0): do not edit by hand
\name{parseresults}
\alias{parseresults}
\title{Parse results from call to occurrencelist endpoint}
\usage{
parseresults(x, ..., removeZeros = removeZeros)
}
\arguments{
\item{x}{A list}

\item{...}{Further args passed on to gbifxmlToDataFrame}

\item{removeZeros}{remove zeros or not}
}
\description{
Parse results from call to occurrencelist endpoint
}
\keyword{internal}

