% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.gbiflist}
\alias{print.gbiflist}
\title{Print summary of gbiflist class}
\usage{
\method{print}{gbiflist}(x, ...)
}
\arguments{
\item{x}{an object of class gbiflist}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print summary of gbiflist class
}

