% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.gbiflist_na}
\alias{print.gbiflist_na}
\title{Print summary of gbiflist_na class}
\usage{
\method{print}{gbiflist_na}(x, ...)
}
\arguments{
\item{x}{an object of class gbiflist_na}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print summary of gbiflist_na class
}

