\name{rank}
\alias{rank}
\title{Get values to be used for rank arguments in GBIF API methods.}
\usage{
  rank()

  rank()
}
\arguments{
  \item{None}{}

  \item{None}{}
}
\description{
  Get values to be used for rank arguments in GBIF API
  methods.

  Get values to be used for rank arguments in GBIF API
  methods.
}
\examples{
\dontrun{
rank()
}
\dontrun{
rank()
}
}

