% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_wkt.R
\name{read_wkt}
\alias{read_wkt}
\title{Check input WKT}
\usage{
read_wkt(wkt)
}
\arguments{
\item{wkt}{(character) A Well Known Text string}
}
\description{
Check input WKT
}
\examples{
wkt <- 'LINESTRING (30 10, 10 30, 40 40)'
read_wkt(wkt)

wkt <- "POLYGON((38.4 -125,40.9 -125,40.9 -121.8,38.4 -121.8,38.4 -125))"
read_wkt(wkt)
}

