% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rgbif-package.r
\name{rgbif-defunct}
\alias{rgbif-defunct}
\title{Defunct functions in rgbif}
\description{
\itemize{
 \item \code{\link{density_spplist}}: service no longer provided
 \item \code{\link{densitylist}}: service no longer provided
 \item \code{\link{gbifdata}}: service no longer provided
 \item \code{\link{gbifmap_dens}}: service no longer provided
 \item \code{\link{gbifmap_list}}: service no longer provided
 \item \code{\link{occurrencedensity}}: service no longer provided
 \item \code{\link{providers}}: service no longer provided
 \item \code{\link{resources}}: service no longer provided
 \item \code{\link{taxoncount}}: service no longer provided
 \item \code{\link{taxonget}}: service no longer provided
 \item \code{\link{taxonsearch}}: service no longer provided
 \item \code{\link{stylegeojson}}: moving this functionality to spocc package, will be removed soon
 \item \code{\link{togeojson}}: moving this functionality to spocc package, will be removed soon
 \item \code{\link{gist}}: moving this functionality to spocc package, will be removed soon
}
}
\details{
The above functions have been removed. See \url{https://github.com/ropensci/rgbif} and poke
around the code if you want to find the old functions in previous versions of the package, or
email Scott at \email{myrmecocystus@gmail.com}
}

