\name{rgbif-deprecated}
\alias{rgbif-deprecated}
\title{Deprecated functions in rgbif}
\description{
  \itemize{ \item \code{\link{density_spplist}}: service no
  longer provided \item \code{\link{densitylist}}: service
  no longer provided \item \code{\link{gbifdata}}: service
  no longer provided \item \code{\link{gbifmap_dens}}:
  service no longer provided \item
  \code{\link{gbifmap_list}}: service no longer provided
  \item \code{\link{is.gbifdens}}: service no longer
  provided \item \code{\link{is.gbiflist}}: service no
  longer provided \item \code{\link{occurrencedensity}}:
  service no longer provided \item \code{\link{providers}}:
  service no longer provided \item \code{\link{resources}}:
  service no longer provided \item
  \code{\link{taxoncount}}: service no longer provided
  \item \code{\link{taxonget}}: service no longer provided
  \item \code{\link{taxonsearch}}: service no longer
  provided }
}

