% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbif-package.r
\docType{package}
\name{rgbif-package}
\alias{rgbif-package}
\alias{rgbif}
\title{Interface to the Global Biodiversity Information Facility API.}
\description{
rgbif: A programmatic interface to the Web Service methods
provided by the Global Biodiversity Information Facility.
}
\section{About}{


This package gives you access to data from GBIF \url{https://www.gbif.org/}
via their API.
}

\section{Documentation for the GBIF API}{

\itemize{
\item summary \url{https://www.gbif.org/developer/summary} - Summary of
the GBIF API
\item registry \url{https://www.gbif.org/developer/registry} - Metadata
on datasets, and contributing organizations
\item species names \url{https://www.gbif.org/developer/species} - Species
names and metadata
\item occurrences \url{https://www.gbif.org/developer/occurrence} -
Occurrences
\item maps \url{https://www.gbif.org/developer/maps} - Maps - these APIs
are not implemented in \pkg{rgbif}, and are meant more for intergration
with web based maps.
}
}

\author{
Scott Chamberlain

Karthik Ram

Dan Mcglinn

Vijay Barve

John Waller
}
