% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbif-package.r
\docType{data}
\name{typestatus}
\alias{typestatus}
\title{Type status options for GBIF searching}
\description{
\itemize{
\item name. Name of type.
\item description. Description of the type.
}
}
\keyword{data}
