% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_name_usage.R
\name{cb_name_usage}
\alias{cb_name_usage}
\title{Search for names on checklistbank}
\usage{
cb_name_usage(
  q = NULL,
  key = "304862",
  id = NULL,
  name = NULL,
  scientificName = NULL,
  authorship = NULL,
  code = NULL,
  rank = NULL,
  status = NULL,
  verbose = NULL,
  superkingdom = NULL,
  kingdom = NULL,
  subkingdom = NULL,
  superphylum = NULL,
  phylum = NULL,
  subphylum = NULL,
  superclass = NULL,
  class = NULL,
  subclass = NULL,
  superorder = NULL,
  order = NULL,
  suborder = NULL,
  superfamily = NULL,
  family = NULL,
  subfamily = NULL,
  tribe = NULL,
  subtribe = NULL,
  genus = NULL,
  subgenus = NULL,
  section = NULL,
  species = NULL,
  user = NULL,
  pwd = NULL,
  curlopts = list()
)
}
\arguments{
\item{q}{(string) General query term.}

\item{key}{(integer) The unique identifier for the resource.}

\item{id}{(string) Identifier of the taxonomic entity.}

\item{name}{(string) The name of the taxonomic entity to query.}

\item{scientificName}{(string) Scientific name of the species or taxonomic group.}

\item{authorship}{(string) Authorship of the scientific name.}

\item{code}{(string) Nomenclatural code (e.g. ICZN, ICBN).}

\item{rank}{(string) Taxonomic rank to filter by (e.g., species, genus).}

\item{status}{(string) Status of the taxon (e.g., valid, synonym).}

\item{verbose}{(boolean) If \code{TRUE}, returns additional information.}

\item{superkingdom}{(string) Taxonomic level above kingdom.}

\item{kingdom}{(string) The kingdom to filter by.}

\item{subkingdom}{(string) The subkingdom to filter by.}

\item{superphylum}{(string) Taxonomic level above phylum.}

\item{phylum}{(string) The phylum to filter by.}

\item{subphylum}{(string) The subphylum to filter by.}

\item{superclass}{(string) Taxonomic level above class.}

\item{class}{(string) The class to filter by.}

\item{subclass}{(string) The subclass to filter by.}

\item{superorder}{(string) Taxonomic level above order.}

\item{order}{(string) The order to filter by.}

\item{suborder}{(string) The suborder to filter by.}

\item{superfamily}{(string) Taxonomic level above family.}

\item{family}{(string) The family to filter by.}

\item{subfamily}{(string) The subfamily to filter by.}

\item{tribe}{(string) Taxonomic tribe to filter by.}

\item{subtribe}{(string) The subtribe to filter by.}

\item{genus}{(string) The genus to filter by.}

\item{subgenus}{(string) The subgenus to filter by.}

\item{section}{(string) The section to filter by within a genus.}

\item{species}{(string) The species to filter by.}

\item{user}{(string) GBIF user name for authentication.}

\item{pwd}{(string) GBIF password for authentication.}

\item{curlopts}{(list) List of named curl options passed on to HttpClient.}
}
\value{
list
}
\description{
Constructs and sends a query to the Taxonomy API based on specified parameters.
}
\examples{
\dontrun{
cb_name_usage("Calopteryx splendens")
}
}
