% Generated by roxygen2 (4.0.0): do not edit by hand
\name{check_wkt}
\alias{check_wkt}
\title{Check input WKT}
\usage{
check_wkt(wkt = NULL)
}
\arguments{
\item{wkt}{A Well Known Text object}
}
\description{
Check input WKT
}
\examples{
check_wkt('POLYGON((30.1 10.1, 10 20, 20 60, 60 60, 30.1 10.1))')
check_wkt('POINT(30.1 10.1)')
check_wkt('LINESTRING(3 4,10 50,20 25)')
}

