% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\title{Title}
\usage{
dataset(
  country = NULL,
  type = NULL,
  identifierType = NULL,
  identifier = NULL,
  machineTagNamespace = NULL,
  machineTagName = NULL,
  machineTagValue = NULL,
  modified = NULL,
  query = NULL,
  limit = NULL,
  start = NULL,
  curlopts = list()
)
}
\arguments{
\item{country}{}

\item{type}{}

\item{identifierType}{}

\item{identifier}{}

\item{machineTagNamespace}{}

\item{machineTagName}{}

\item{machineTagValue}{}

\item{modified}{}

\item{query}{}

\item{limit}{}

\item{start}{}

\item{curlopts}{}
}
\description{
Title
}
\examples{
\dontrun {

}
}
