% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_doi.R
\name{dataset_doi}
\alias{dataset_doi}
\title{Get a GBIF dataset from a doi}
\usage{
dataset_doi(doi = NULL, limit = 20, start = NULL, curlopts = list())
}
\arguments{
\item{doi}{the doi of the dataset you wish to lookup.}

\item{limit}{Controls the number of results in the page.}

\item{start}{Determines the offset for the search results.}

\item{curlopts}{options passed on to \link[crul:HttpClient]{crul::HttpClient}.}
}
\value{
A \code{list}.
}
\description{
Get a GBIF dataset from a doi
}
\details{
This function allows for dataset lookup using a doi. Be aware that
some doi have more than one dataset associated with them.
}
\examples{
\dontrun {
dataset_doi("10.15468/igasai")
}
}
