% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dataset_metrics}
\alias{dataset_metrics}
\title{Geta details on a dataset.}
\usage{
dataset_metrics(uuid, callopts = list())
}
\arguments{
\item{uuid}{(character) One or more dataset UUIDs. See examples.}

\item{callopts}{Pass on options to GET.}
}
\description{
Geta details on a dataset.
}
\examples{
\dontrun{
dataset_metrics(uuid='3f8a1297-3259-4700-91fc-acc4170b27ce')
dataset_metrics(uuid='66dd0960-2d7d-46ee-a491-87b9adcfe7b1')
dataset_metrics(uuid=c('3f8a1297-3259-4700-91fc-acc4170b27ce',
   '66dd0960-2d7d-46ee-a491-87b9adcfe7b1'))

library("httr")
dataset_metrics(uuid='66dd0960-2d7d-46ee-a491-87b9adcfe7b1', verbose())
}
}
\references{
\url{http://www.gbif.org/developer/summary}
}

