% Generated by roxygen2 (4.0.2): do not edit by hand
\name{density_spplist}
\alias{density_spplist}
\title{The density web service provides access to records showing the density
   of occurrence records from the GBIF Network by one-degree cell.}
\usage{
density_spplist(...)
}
\description{
This function is defunct
}
\keyword{internal}

