% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_user_dl.R
\name{dl_exists}
\alias{dl_exists}
\title{Does a download exist already on the GBIF's user account remotely?}
\usage{
dl_exists(pred, preds)
}
\arguments{
\item{pred}{a download request predicate, the output from
\code{\link[=occ_download_prep]{occ_download_prep()}}}

\item{preds}{output from \code{\link[=dl_predicates]{dl_predicates()}}}
}
\value{
boolean
}
\description{
Does a download exist already on the GBIF's user account remotely?
}
\details{
we convert predicates to JSON, then compare JSON blobs to one
another
}
\examples{
\dontrun{
preds <- dl_predicates(dl_user())

dprep <- occ_download_prep("catalogNumber = Bird.27847588", "year = 1978",
  "month = 5")
dprep <- occ_download_prep(pred("catalogNumber", "Bird.27847588"), pred("year", 1978))
dl_exists(pred = dprep, preds)

dprep <- occ_download_prep(pred("catalogNumber", "Bird.27847588"))
dl_exists(pred = dprep, preds)
}
}
\keyword{internal}
