% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_user_dl.R
\name{dl_predicates}
\alias{dl_predicates}
\title{Fetch all user download predicates}
\usage{
dl_predicates(user_df)
}
\arguments{
\item{user_df}{output from \code{\link[=dl_user]{dl_user()}}}
}
\description{
Fetch all user download predicates
}
\examples{
\dontrun{
x <- dl_user()
res <- dl_predicates(user_df = x)
res
}
}
\keyword{internal}
