% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_user_dl.R
\name{dl_user}
\alias{dl_user}
\title{get all user downloads}
\usage{
dl_user(user = NULL, pwd = NULL, curlopts = list())
}
\description{
get all user downloads
}
\examples{
x <- dl_user()
x
}
\keyword{internal}
