% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_queue.R
\docType{data}
\name{DownReq}
\alias{DownReq}
\title{Download request}
\arguments{
\item{x}{either a lazy object with an object of class \code{occ_download}, or an
object of class \code{occ_download_prep}}
}
\description{
Download request
}
\details{
\strong{Methods}
\describe{
\item{\code{new(...)}}{
initialize request
}
\item{\code{run(...)}}{
execute http request
}
\item{\code{status(...)}}{
check http request status
}
}
}
\examples{
\dontrun{
# res <- DownReq$new(occ_download('taxonKey = 3119195', "year = 1991"))
# res
# res$req
# res$run()
# (requests <- GbifQueue$new())
# res$run(keep_track = TRUE)
# requests
# res$status()

# prepared query
res <- DownReq$new(occ_download_prep("basisOfRecord = LITERATURE"))
res
res$run()
res
res$status()
res$result
}
}
\keyword{internal}
