% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation.r
\name{elevation}
\alias{elevation}
\title{Get elevation for lat/long points from a data.frame or list of points.}
\usage{
elevation(input = NULL, latitude = NULL, longitude = NULL,
  latlong = NULL, username = Sys.getenv("GEONAMES_USER"), ...)
}
\arguments{
\item{input}{A data.frame of lat/long data. There must be columns
decimalLatitude and decimalLongitude.}

\item{latitude}{A vector of latitude's. Must be the same length as the
longitude vector.}

\item{longitude}{A vector of longitude's. Must be the same length as
the latitude vector.}

\item{latlong}{A vector of lat/long pairs. See examples.}

\item{username}{(character) Required. An GeoNames user name. See Details.}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}
see \code{\link[curl:curl_options]{curl::curl_options()}} for curl options}
}
\value{
A new column named elevation in the supplied data.frame or a vector
with elevation of each location in meters.
}
\description{
Uses the GeoNames web service
}
\details{
To get a GeoNames user name, register for an account at
http://www.geonames.org/login - then you can enable your account for the
GeoNames webservice on your account page. Once you are enabled to use
the webservice, you can pass in your username to the \code{username}
parameter. Better yet, store your username in your \code{.Renviron} file, or
similar (e.g., .zshrc or .bash_profile files) and read it in via
\code{Sys.getenv()} as in the examples below. By default we do
\code{Sys.getenv("GEONAMES_USER")} for the \code{username} parameter.
}
\examples{
\dontrun{
user <- Sys.getenv("GEONAMES_USER")

occ_key <- name_suggest('Puma concolor')$key[1]
dat <- occ_search(taxonKey = occ_key, return = 'data', limit = 300,
  hasCoordinate = TRUE)
head( elevation(dat, username = user) )

# Pass in a vector of lat's and a vector of long's
elevation(latitude = dat$decimalLatitude[1:10],
  longitude = dat$decimalLongitude[1:10],
  username = user)

# Pass in lat/long pairs in a single vector
pairs <- list(c(31.8496,-110.576060), c(29.15503,-103.59828))
elevation(latlong=pairs, username = user)

# Pass on curl options
pairs <- list(c(31.8496,-110.576060), c(29.15503,-103.59828))
elevation(latlong=pairs, username = user, verbose = TRUE)
}
}
\references{
GeoNames http://www.geonames.org/export/web-services.html
}
