% Generated by roxygen2 (4.0.1): do not edit by hand
\name{elevation}
\alias{elevation}
\title{Get elevation for lat/long points from a data.frame or list of points.}
\usage{
elevation(input = NULL, latitude = NULL, longitude = NULL,
  latlong = NULL, callopts = list())
}
\arguments{
\item{input}{A data.frame of lat/long data. There must be columns decimalLatitude and
decimalLongitude.}

\item{latitude}{A vector of latitude's. Must be the same length as the longitude
vector.}

\item{longitude}{A vector of longitude's. Must be the same length as the latitude
vector.}

\item{latlong}{A vector of lat/long pairs. See examples.}

\item{callopts}{Options passed on to \code{httr::GET}, like curl options for debugging.}
}
\value{
A new column named elevation in the supplied data.frame or a vector with elevation of
each location in meters.
}
\description{
Get elevation for lat/long points from a data.frame or list of points.
}
\examples{
\dontrun{
key <- name_suggest('Puma concolor')$key[1]
dat <- occ_search(taxonKey=key, return='data', limit=300, hasCoordinate=TRUE)
elevation(dat)

# Pass in a vector of lat's and a vector of long's
elevation(latitude=dat$decimalLatitude, longitude=dat$decimalLongitude)

# Pass in lat/long pairs in a single vector
pairs <- list(c(31.8496,-110.576060), c(29.15503,-103.59828))
elevation(latlong=pairs)
}
}
\references{
Uses the Google Elevation API at the following link
\url{https://developers.google.com/maps/documentation/elevation/}
}

