% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumeration.R
\name{enumeration}
\alias{enumeration}
\alias{enumeration_country}
\title{Enumerations.}
\usage{
enumeration(x = NULL, ...)

enumeration_country(...)
}
\arguments{
\item{x}{A given enumeration.}

\item{...}{Further named parameters, such as \code{query}, \code{path}, etc, passed on to
\code{\link[httr]{modify_url}} within \code{\link[httr]{GET}} call. Unnamed parameters will
be combined with \code{\link[httr]{config}}.}
}
\value{
\code{enumeration} returns a character vector, while
\code{enumeration_country} returns a data.frame.
}
\description{
Many parts of the GBIF API make use of enumerations, i.e. controlled
vocabularies for specific topics - and are available via these functions
}
\examples{
\dontrun{
# basic enumeration
enumeration()
enumeration("NameType")
enumeration("MetadataType")
enumeration("TypeStatus")

# country enumeration
enumeration_country()

# curl options
library("httr")
enumeration(config = verbose())
}
}

