% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbif_issues.R
\name{gbif_issues}
\alias{gbif_issues}
\title{Table of GBIF issues, with codes used in data output, full issue name,
and descriptions.}
\source{
\url{http://gbif.github.io/gbif-api/apidocs/org/gbif/api/vocabulary/OccurrenceIssue.html}
}
\usage{
gbif_issues()
}
\description{
Table has the following fields:
}
\details{
\itemize{
\item code. Code for issue, making viewing data easier.
\item issue. Full name of the issue.
\item description. Description of the issue.
}
}
