\name{gbifdatause}
\alias{gbifdatause}
\title{gbifdatause - description}
\usage{
  gbifdatause(url = "http://data.gbif.org/tutorial/datasharingagreement",
    ..., curl = getCurlHandle())
}
\arguments{
  \item{None}{}
}
\description{
  gbifdatause - description
}
\examples{
\dontrun{
gbifdatause()
}
}

