\name{gbifdatause}
\alias{gbifdatause}
\title{gbifdatause - description}
\usage{
  gbifdatause(url = "http://data.gbif.org/tutorial/datasharingagreement")
}
\arguments{
  \item{url}{Base url for call, don't change.}
}
\description{
  gbifdatause - description
}
\examples{
\dontrun{
gbifdatause()
}
}

