\name{gbifmap}
\alias{gbifmap}
\title{Plot a class of gbiflist, gbifdensity}
\usage{
  gbifmap(input, ...)
}
\arguments{
  \item{input}{Input data.frame}

  \item{...}{Additional parameters passed on to
  gbifmap.gbiflist or gbifmap.gbifdens}
}
\description{
  Plot a class of gbiflist, gbifdensity
}

