% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gbifparser_verbatim}
\alias{gbifparser_verbatim}
\title{Parser for gbif data}
\usage{
gbifparser_verbatim(input, fields = "minimal")
}
\arguments{
\item{input}{A list}

\item{fields}{(character) Default ('minimal') will return just taxon name, key, decimalLatitude, and
decimalLongitute. 'all' returns all fields. Or specify each field you want returned by name, e.g.
fields = c('name','decimalLatitude','altitude').}
}
\description{
Parser for gbif data
}
\keyword{internal}

