% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_search.R
\name{lit_search}
\alias{lit_search}
\alias{lit_all_gbif}
\alias{lit_all_dataset}
\alias{lit_count}
\title{Search for literature that cites GBIF mediated data}
\usage{
lit_search(
  q = NULL,
  countriesOfResearcher = NULL,
  countriesOfCoverage = NULL,
  literatureType = NULL,
  relevance = NULL,
  year = NULL,
  topics = NULL,
  gbifDatasetKey = NULL,
  publishingOrganizationKey = NULL,
  peerReview = NULL,
  openAccess = NULL,
  gbifDownloadKey = NULL,
  doi = NULL,
  source = NULL,
  publisher = NULL,
  flatten = TRUE,
  start = 0,
  limit = 500,
  curlopts = list()
)

lit_all_gbif(nrows = NULL, peerReview = TRUE, flatten = TRUE, step = 1000)

lit_all_dataset(
  gbifDatasetKey = NULL,
  peerReview = NULL,
  nrows = NULL,
  step = 1000,
  flatten = TRUE
)

lit_count(...)
}
\arguments{
\item{q}{(character) Simple full text search parameter. The value for this
parameter can be a simple word or a phrase. Wildcards are not supported.}

\item{countriesOfResearcher}{(character) Country of institution with which
author is affiliated, e.g. DK (for Denmark). Country codes are
listed in enumeration_country().}

\item{countriesOfCoverage}{(character) Country of focus of study,
e.g. BR (for Brazil). Country codes are listed in enumeration_country().}

\item{literatureType}{(character) Type of literature ("JOURNAL",
"BOOK_SECTION", "WORKING_PAPER", "REPORT", "GENERIC", "THESIS", "CONFERENCE_PROCEEDINGS",
"WEB_PAGE").}

\item{relevance}{(character) How is the publication relate to GBIF. See details
("GBIF_USED", "GBIF_MENTIONED", "GBIF_PUBLISHED", "GBIF_CITED", "GBIF_CITED",
"GBIF_PUBLISHED", "GBIF_ACKNOWLEDGED", "GBIF_AUTHOR").}

\item{year}{(integer) Year of publication.}

\item{topics}{(character) Topic of publication.}

\item{gbifDatasetKey}{(character) GBIF dataset uuid referenced in publication.}

\item{publishingOrganizationKey}{(character) Publisher uuid whose dataset is
referenced in publication.}

\item{peerReview}{(logical) Has publication undergone peer-review?}

\item{openAccess}{(logical) Is publication Open Access?}

\item{gbifDownloadKey}{(character) Download referenced in publication.}

\item{doi}{(character) Digital Object Identifier (DOI).}

\item{source}{(character) Journal of publication.}

\item{publisher}{(character) Publisher of journal.}

\item{flatten}{(logical) should any lists in the resulting data be flattened
into comma-seperated strings?}

\item{start}{Record number to start at. Default: 0. Use in combination
with \code{limit} to page through results.}

\item{limit}{Number of records to return. Default: 100. Maximum: 1000.}
}
\value{
A named list with two values: \code{$data} and \code{$meta}. \code{$data} is
a \code{data.frame} of literature references.
}
\description{
Search for literature that cites GBIF mediated data
}
\details{
This function enables you to search for literature indexed by GBIF,
including peer-reviewed papers, citing GBIF datasets and downloads.
The literature API powers the
\href{https://www.gbif.org/resource/search?contentType=literature}{literature search}
on GBIF.

The GBIF Secretariat maintains an ongoing
\href{https://www.gbif.org/literature-tracking}{literature tracking programme},
which identifies research uses and citations of biodiversity information
accessed through GBIF’s global infrastructure.

In the literature database, \strong{relevance} refers to how publications relate
to GBIF following these definitions:
\itemize{
\item GBIF_USED : makes substantive use of data in a quantitative analysis (e.g. ecological niche modelling)
\item GBIF_CITED : cites a qualitative fact derived in data (e.g. a given species is found in a given country)
\item GBIF_DISCUSSED : discusses GBIF as an infrastructure or the use of data
\item GBIF_PRIMARY : GBIF is the primary source of data (no longer applied)
\item GBIF_ACKNOWLEDGED : acknowledges GBIF (but doesn't use or cite data)
\item GBIF_PUBLISHED : describes or talks about data published to GBIF
\item GBIF_AUTHOR : authored by GBIF staff
\item GBIF_MENTIONED : unspecifically mentions GBIF or the GBIF portal
\item GBIF_FUNDED : funded by GBIF or a GBIF-managed funding programme
}

The following arguments can take multiple values:
\itemize{
\item relevance
\item countriesOfResearcher
\item countriesOfCoverage
\item literatureType
\item topics
\item gbifDatasetKey
\item publishingOrganizationKey
\item gbifDownloadKey
\item doi
\item source
\item publisher
}

If \code{flatten=TRUE}, then \strong{data} will be returned as flat
data.frame with no complex column types (i.e. no lists or data.frames).

\code{lit_all_gbif()} is a convenience wrapper, which will return all
peer-reviewed literature that cites or uses GBIF mediated data in a single
\code{data.frame}.

\code{lit_all_dataset()} is a convenience wrapper, which will return
literature citations for a given GBIF dataset uuid. This corresponds to what
currently appears on \href{https://www.gbif.org/dataset/50c9509d-22c7-4a22-a47d-8c48425ef4a7}{dataset pages}
for "citations".

\code{lit_count()} is a convenience wrapper, which will return the number of
literature references for a certain \code{lit_search()} query.
}
\examples{
\dontrun{
 # https://www.gbif.org/resource/search?contentType=literature
 
 lit_search(q="bats")$data 
 lit_search(gbifDatasetKey="50c9509d-22c7-4a22-a47d-8c48425ef4a7")
 lit_search(year=2020)
 lit_search(year="2011,2020") # year ranges
 lit_search(relevance=c("GBIF_CITED","GBIF_USED")) # multiple values
 lit_search(relevance=c("GBIF_USED","GBIF_CITED"), 
 topics=c("EVOLUTION","PHYLOGENETICS")

}


}
