% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_backbone_checklist.R
\name{name_backbone_checklist}
\alias{name_backbone_checklist}
\title{Title}
\usage{
name_backbone_checklist(
  name_data = NULL,
  strict = FALSE,
  start = NULL,
  limit = 100,
  curlopts = list(),
  verbose = FALSE
)
}
\arguments{
\item{name_data}{}

\item{strict}{}

\item{start}{}

\item{limit}{}

\item{curlopts}{}

\item{verbose}{}
}
\value{

}
\description{
Title
}
\examples{

}
