% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_issues.R
\name{name_issues}
\alias{name_issues}
\title{Parse and examine further GBIF name issues on a dataset.}
\usage{
name_issues(.data, ..., mutate = NULL)
}
\arguments{
\item{.data}{Output from a call to \code{\link[=name_usage]{name_usage()}}, but only if
\code{return="all"}, or \code{return="data"}, otherwise function stops with error.}

\item{...}{Named parameters to only get back (e.g. bbmn), or to
remove (e.g. -bbmn).}

\item{mutate}{(character) One of:
\itemize{
\item \code{split} Split issues into new columns.
\item \code{expand} Expand issue abbreviated codes into descriptive names.
for downloads datasets, this is not super useful since the
issues come to you as expanded already.
\item \code{split_expand} Split into new columns, and expand issue names.
}

For split and split_expand, values in cells become y ("yes") or n ("no")}
}
\description{
Parse and examine further GBIF name issues on a dataset.
}
\references{
\url{https://gbif.github.io/gbif-api/apidocs/org/gbif/api/vocabulary/NameUsageIssue.html}
}
