% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_issues.R
\name{name_issues}
\alias{name_issues}
\title{Table of GBIF name usage issues, with codes used in data output, full issue name,
and descriptions.}
\source{
\url{https://gbif.github.io/gbif-api/apidocs/org/gbif/api/vocabulary/NameUsageIssue.html}
}
\usage{
name_issues()
}
\description{
Table has the following fields:
}
\details{
\itemize{
\item code. Code for issue, making viewing data easier.
\item issue. Full name of the issue.
\item description. Description of the issue.
}
}
\seealso{
\code{\link[=gbif_issues]{gbif_issues()}}
}
