% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_count.R
\name{occ_count}
\alias{occ_count}
\title{Get number of occurrence records.}
\usage{
occ_count(..., curlopts = list())
}
\arguments{
\item{...}{parameters passed to \code{occ_search()}.}

\item{curlopts}{(list) curl options.}
}
\value{
The occurrence count of the \code{occ_search()} query.
}
\description{
Get number of occurrence records.
}
\details{
\code{occ_count()} is a short convenience wrapper for
\code{occ_search(limit=0,occurrenceStatus=NULL)$meta$count}.

The current version of \code{occ_count()} uses a different GBIF API endpoint than
than previous versions. This change greatly improves the usability of
\code{occ_count()}. Legacy parameters \code{georeferenced}, \code{type}, \code{date}, \code{to},
\code{from} are no longer supported and not guaranteed to work correctly. See
\code{occ_counts_country}, \code{occ_counts_pub_country()}, \code{occ_counts_year()},
\code{occ_counts_year}, and  \code{occ_counts_basis_of_record()} for replacements of
counts available via the \code{type} parameter.

Certain \code{occ_search()} parameters, such as \code{facet}, will be ignored since
they do not make sense in the context of a total count. Multiple values of
the type \code{c("a","b")} will give an error, but \code{"a;b"} will work.
}
\examples{
\dontrun{
# total occurrences mediated by GBIF
occ_count() # should be > 2 billion! 

# number of plant occurrences
occ_count(kingdomKey=name_backbone("Plantea")$usageKey) 
occ_count(scientificName = 'Ursus americanus')


occ_count(country="DK") # found in Denmark 
occ_count(country="DK;US") # found in Denmark and United States
occ_count(publishingCountry="US") # published by the United States
# number of repatriated eBird records in India
occ_count(repatriated = TRUE,country="IN") 
 
occ_count(taxonKey=212) # number of bird occurrences
# between years 1800-1900
occ_count(basisOfRecord="PRESERVED_SPECIMEN", year="1800,1900") 
occ_count(recordedBy="John Waller") # recorded by John Waller
occ_count(decimalLatitude=0, decimalLongitude=0) # exactly on 0,0

# close to a known iso2 centroid
occ_count(distanceFromCentroidInMeters="0,2000") 
# close to a known iso2 centroid in Sweden
occ_count(distanceFromCentroidInMeters="0,2000",country="SE") 

occ_count(hasCoordinate=TRUE) # with coordinates
occ_count(protocol = "DIGIR") # published using DIGIR format
occ_count(mediaType = 'StillImage') # with images

# number of occurrences iucn status "critically endangered"
occ_count(iucnRedListCategory="CR") 
occ_count(verbatimScientificName="Calopteryx splendens;Calopteryx virgo")
occ_count(
geometry="POLYGON((24.70938 48.9221,24.71056 48.92175,24.71107
 48.92296,24.71002 48.92318,24.70938 48.9221))")

}
}
