% Generated by roxygen2 (4.0.2): do not edit by hand
\name{occ_download_cancel}
\alias{occ_download_cancel}
\title{Cancel a download process.}
\usage{
occ_download_cancel(key, user = getOption("gbif_user"),
  pwd = getOption("gbif_pwd"), ...)
}
\arguments{
\item{key}{A key generated from a request, like that from \code{occ_download}. Required.}

\item{user}{(character) User name within GBIF's website. Required.}

\item{pwd}{(character) User password within GBIF's website. Required.}

\item{...}{Further args passed to \code{\link[httr]{DELETE}}}
}
\description{
Cancel a download process.
}
\examples{
\donttest{
occ_download_cancel(key="0003984-140910143529206")
}
}

