% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_dataset_activity.R
\name{occ_download_dataset_activity}
\alias{occ_download_dataset_activity}
\title{Lists the downloads activity of a dataset}
\usage{
occ_download_dataset_activity(dataset, curlopts = list())
}
\arguments{
\item{dataset}{(character) A dataset key}

\item{curlopts}{list of named curl options passed on to
\code{\link[crul]{HttpClient}}. see \code{\link[curl]{curl_options}}
for curl options}
}
\value{
a list with slots of offset, limit, endOfRecords, count, and results.
results has an array of the downloads for the dataset
}
\description{
Lists the downloads activity of a dataset
}
\note{
see \link{downloads} for an overview of GBIF downloads methods
}
\examples{
\dontrun{
res <- occ_download_dataset_activity("7f2edc10-f762-11e1-a439-00145eb45e9a")
res
res$count
}
}
