% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_datasets.R
\name{occ_download_datasets}
\alias{occ_download_datasets}
\title{List datasets for a download}
\usage{
occ_download_datasets(key, curlopts = list())
}
\arguments{
\item{key}{A key generated from a request, like that from
\code{occ_download}}

\item{curlopts}{list of named curl options passed on to
\code{\link[crul]{HttpClient}}. see \code{\link[curl]{curl_options}}
for curl options}
}
\value{
a list with slots of offset, limit, endOfRecords, count, and results.
results has an array of the datasets
}
\description{
List datasets for a download
}
\examples{
\dontrun{
occ_download_datasets(key="0003983-140910143529206")
}
}
