% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_describe.R
\name{occ_download_describe}
\alias{occ_download_describe}
\title{Describes the fields available in GBIF downloads}
\usage{
occ_download_describe(x = "dwca")
}
\arguments{
\item{x}{a character string (default: "dwca"). Accepted values:
"simpleCsv", "simpleAvro", "simpleParquet","speciesList", "sql".}
}
\value{
a list.
}
\description{
Describes the fields available in GBIF downloads
}
\details{
The function returns a list with the fields available in GBIF downloads. It
is considered experimental by GBIF, so the output might change in the future.
}
\examples{
\dontrun{
occ_download_describe("dwca")$verbatimFields
occ_download_describe("dwca")$verbatimExtensions
occ_download_describe("simpleCsv")$fields
occ_download_describe("sql")$fields
}
}
