% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_doi.R
\name{occ_download_doi}
\alias{occ_download_doi}
\title{Get download meta data from a doi}
\usage{
occ_download_doi(doi = NULL)
}
\arguments{
\item{doi}{(character) the doi of the download you want to get metadata for.}
}
\value{
a list.
}
\description{
Get download meta data from a doi
}
\examples{
\dontrun{
occ_download_doi("10.15468/dl.zdfkkf")

occ_download_doi("10.15468/dl.zdfkkf")$key \%>\%
occ_download_get() \%>\%
occ_download_import() 

}

}
