% Generated by roxygen2 (4.0.2): do not edit by hand
\name{occ_download_import}
\alias{occ_download_import}
\title{Import a downloaded file from GBIF.}
\usage{
occ_download_import(x = NULL, key = NULL, path = "~/")
}
\arguments{
\item{x}{The output of a call to \code{occ_download_get}}

\item{key}{A key generated from a request, like that from \code{occ_download}}

\item{path}{Path to write file to. Default: "~/" your home directory, with a \code{.zip}
appended to the end.}
}
\description{
Import a downloaded file from GBIF.
}
\details{
You can provide either x as input, or both key and path.
}
\examples{
\donttest{
res <- occ_download_get(key="0000066-140928181241064")
occ_download_import(x=res)

occ_download_get(key="0000066-140928181241064") \%>\% occ_download_import
}
}

