% Generated by roxygen2 (4.0.2): do not edit by hand
\name{occ_download_list}
\alias{occ_download_list}
\title{Lists the downloads created by a user.}
\usage{
occ_download_list(user = getOption("gbif_user"),
  pwd = getOption("gbif_pwd"), ...)
}
\arguments{
\item{user}{A user name, look at option "gbif_user" first}

\item{pwd}{Your password, look at option "gbif_pwd" first}

\item{...}{Further args passed to \code{\link[httr]{GET}}}
}
\description{
Lists the downloads created by a user.
}
\examples{
\donttest{
occ_download_list(user="sckott")
}
}

