% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_list.R
\name{occ_download_list}
\alias{occ_download_list}
\title{Lists the downloads created by a user.}
\usage{
occ_download_list(user = getOption("gbif_user"),
  pwd = getOption("gbif_pwd"), limit = 20, start = 0, ...)
}
\arguments{
\item{user}{A user name, look at option "gbif_user" first}

\item{pwd}{Your password, look at option "gbif_pwd" first}

\item{limit}{Number of records to return. Default: 20}

\item{start}{Record number to start at. Default: 0}

\item{...}{Further args passed to \code{\link[httr]{GET}}}
}
\description{
Lists the downloads created by a user.
}
\examples{
\dontrun{
occ_download_list(user="sckott")
occ_download_list(user="sckott", limit = 5)
occ_download_list(user="sckott", start = 21)
}
}

