% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_organizations.R
\name{occ_download_organizations}
\alias{occ_download_organizations}
\title{List organizations for a download}
\usage{
occ_download_organizations(
  key,
  organizationTitle = NULL,
  sortBy = NULL,
  sortOrder = NULL,
  limit = 20,
  start = 0,
  curlopts = list(http_version = 2)
)
}
\arguments{
\item{key}{A key generated from a request, like that from \code{\link[=occ_download]{occ_download()}}}

\item{organizationTitle}{(character) Organization title filter. Optional.}

\item{sortBy}{(character) Sort field. One of \code{ORGANIZATION_TITLE},
\code{COUNTRY_CODE} or \code{RECORD_COUNT}. Optional.}

\item{sortOrder}{(character) Sort order. One of \code{ASC} or \code{DESC}. Optional.}

\item{limit}{(integer/numeric) Number of records to return. Default: 20,
Max: 1000}

\item{start}{(integer/numeric) Record number to start at. Default: 0}

\item{curlopts}{list of named curl options passed on to
\code{\link[crul]{HttpClient}}. see \code{curl::curl_options}
for curl options}
}
\value{
a list with two slots:
\itemize{
\item meta: a single row data.frame with columns: \code{offset}, \code{limit},
\code{endofrecords}, \code{count}
\item results: a tibble with the results, with columns: \code{downloadKey},
\code{organizationKey}, \code{organizationTitle}, \code{numberRecords},
\code{publishingCountryCode}
}
}
\description{
List organizations for a download
}
\note{
see \code{\link[=downloads]{downloads}} for an overview of GBIF downloads methods
}
\examples{
\dontrun{
occ_download_organizations(key="0024953-260519110011954")
occ_download_organizations(key="0024953-260519110011954", limit = 3)
occ_download_organizations(
  key="0024953-260519110011954",
  organizationTitle = "University of Alaska Museum of the North"
)
occ_download_organizations(key="0024953-260519110011954", sortBy = "RECORD_COUNT",
  sortOrder = "DESC")
}
}
