% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_sql.R
\name{occ_download_sql}
\alias{occ_download_sql}
\title{Download occurrence data using a SQL query}
\usage{
occ_download_sql(
  q = NULL,
  format = "SQL_TSV_ZIP",
  user = NULL,
  pwd = NULL,
  email = NULL,
  curlopts = list()
)
}
\arguments{
\item{q}{sql query}

\item{format}{only "SQL_TSV_ZIP" is supported right now}

\item{user}{your GBIF user name}

\item{pwd}{your GBIF password}

\item{email}{your email address}

\item{curlopts}{list of curl options}
}
\value{
an object of class 'occ_download_sql'
}
\description{
Download occurrence data using a SQL query
}
\examples{
/dontrun{

}

}
