% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_stats.R
\name{occ_download_stats_dataset}
\alias{occ_download_stats_dataset}
\title{Downloads by dataset}
\usage{
occ_download_stats_dataset(
  from = NULL,
  to = NULL,
  publishingCountry = NULL,
  datasetKey = NULL,
  publishingOrgKey = NULL,
  curlopts = list(http_version = 2)
)
}
\arguments{
\item{from}{(character) Start date in format \code{YYYY-MM}. Optional.}

\item{to}{(character) End date in format \code{YYYY-MM}. Optional.}

\item{publishingCountry}{(character) ISO 2-letter country code. Optional.}

\item{datasetKey}{(character) Dataset UUID. Optional.}

\item{publishingOrgKey}{(character) Publishing organization UUID. Optional.}

\item{curlopts}{list of named curl options passed on to \link[crul:HttpClient]{crul::HttpClient}.
See \link[curl:curl_options]{curl::curl_options} for curl options}
}
\value{
A tibble with columns: \code{year}, \code{month}, \code{number_downloads}. Each row
represents the total number of downloads for a given year and month.
}
\description{
Summarize downloads by dataset.
}
\note{
see \link{downloads} for an overview of GBIF downloads methods
}
\examples{
\dontrun{
# Get downloads by dataset
occ_download_stats_dataset()

# Filter by date range and publishing country
occ_download_stats_dataset(from = "2023-01", publishingCountry = "US")

# Filter by publishing organization (e.g., iNaturalist)
occ_download_stats_dataset(publishingOrgKey = "28eb1a3f-1c15-4a95-931a-4af90ecb574d")
}
}
\seealso{
Other downloads: 
\code{\link{download_predicate_dsl}},
\code{\link{occ_download_cached}()},
\code{\link{occ_download_cancel}()},
\code{\link{occ_download_dataset_activity}()},
\code{\link{occ_download_datasets}()},
\code{\link{occ_download_get}()},
\code{\link{occ_download_import}()},
\code{\link{occ_download_list}()},
\code{\link{occ_download_meta}()},
\code{\link{occ_download_queue}()},
\code{\link{occ_download_stats_dataset_records}()},
\code{\link{occ_download_stats_export}()},
\code{\link{occ_download_stats_source}()},
\code{\link{occ_download_stats_user_country}()},
\code{\link{occ_download_stats}()},
\code{\link{occ_download_wait}()},
\code{\link{occ_download}()}
}
\concept{downloads}
