% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_stats.R
\name{occ_download_stats_source}
\alias{occ_download_stats_source}
\title{Downloads by source}
\usage{
occ_download_stats_source(
  from = NULL,
  to = NULL,
  source = NULL,
  curlopts = list(http_version = 2)
)
}
\arguments{
\item{from}{(character) Start date in format \code{YYYY-MM}. Optional.}

\item{to}{(character) End date in format \code{YYYY-MM}. Optional.}

\item{source}{(character) Restrict to a particular source (e.g., "rgbif", "pygbif"). Optional.}

\item{curlopts}{list of named curl options passed on to \link[crul:HttpClient]{crul::HttpClient}.
See \link[curl:curl_options]{curl::curl_options} for curl options}
}
\value{
A tibble with columns: \code{year}, \code{month}, \code{number_downloads}. Each row
represents the total number of downloads for a given year and month.
}
\description{
Summarize downloads by source.
}
\note{
see \link{downloads} for an overview of GBIF downloads methods
}
\examples{
\dontrun{
# Get downloads by source
occ_download_stats_source()

# Filter by date range
occ_download_stats_source(from = "2023-01", to = "2023-12")

# Filter by source
occ_download_stats_source(source = "pygbif")
}
}
\seealso{
Other downloads: 
\code{\link{download_predicate_dsl}},
\code{\link{occ_download_cached}()},
\code{\link{occ_download_cancel}()},
\code{\link{occ_download_dataset_activity}()},
\code{\link{occ_download_datasets}()},
\code{\link{occ_download_get}()},
\code{\link{occ_download_import}()},
\code{\link{occ_download_list}()},
\code{\link{occ_download_meta}()},
\code{\link{occ_download_queue}()},
\code{\link{occ_download_stats_dataset_records}()},
\code{\link{occ_download_stats_dataset}()},
\code{\link{occ_download_stats_export}()},
\code{\link{occ_download_stats_user_country}()},
\code{\link{occ_download_stats}()},
\code{\link{occ_download_wait}()},
\code{\link{occ_download}()}
}
\concept{downloads}
