% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_download_stats.R
\name{occ_download_stats_user_country}
\alias{occ_download_stats_user_country}
\title{Downloads by user country}
\usage{
occ_download_stats_user_country(
  from = NULL,
  to = NULL,
  userCountry = NULL,
  publishingCountry = NULL,
  curlopts = list(http_version = 2)
)
}
\arguments{
\item{from}{(character) Start date in format \code{YYYY-MM}. Optional.}

\item{to}{(character) End date in format \code{YYYY-MM}. Optional.}

\item{userCountry}{(character) ISO 2-letter country code. Optional.}

\item{publishingCountry}{(character) ISO 2-letter country code. Optional.}

\item{curlopts}{list of named curl options passed on to \link[crul:HttpClient]{crul::HttpClient}.
See \link[curl:curl_options]{curl::curl_options} for curl options}
}
\value{
A tibble with download counts by user country and month
}
\description{
Summarizes downloads by month, grouped by the user's country code.
}
\note{
see \link{downloads} for an overview of GBIF downloads methods
}
\examples{
\dontrun{
# Get run with no args to get monthly download counts for all of GBIF 
occ_download_stats_user_country()

# Filter by date range
occ_download_stats_user_country(from = "2023-01", to = "2023-12")

# Filter by user country
occ_download_stats_user_country(userCountry = "US")
}
}
\seealso{
Other downloads: 
\code{\link{download_predicate_dsl}},
\code{\link{occ_download_cached}()},
\code{\link{occ_download_cancel}()},
\code{\link{occ_download_dataset_activity}()},
\code{\link{occ_download_datasets}()},
\code{\link{occ_download_get}()},
\code{\link{occ_download_import}()},
\code{\link{occ_download_list}()},
\code{\link{occ_download_meta}()},
\code{\link{occ_download_queue}()},
\code{\link{occ_download_stats_dataset_records}()},
\code{\link{occ_download_stats_dataset}()},
\code{\link{occ_download_stats_export}()},
\code{\link{occ_download_stats_source}()},
\code{\link{occ_download_stats}()},
\code{\link{occ_download_wait}()},
\code{\link{occ_download}()}
}
\concept{downloads}
