% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbif-package.r
\docType{data}
\name{occ_fields}
\alias{occ_fields}
\title{Vector of fields in the output for the function \code{\link[=occ_search]{occ_search()}}}
\description{
These fields can be specified in the \code{fields} parameer in the
\code{\link[=occ_search]{occ_search()}} function.
}
\keyword{data}
