% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_issues_lookup.r
\name{occ_issues_lookup}
\alias{occ_issues_lookup}
\title{Lookup occurrence issue definitions and short codes}
\usage{
occ_issues_lookup(issue = NULL, code = NULL)
}
\arguments{
\item{issue}{Full name of issue, e.g, CONTINENT_COUNTRY_MISMATCH}

\item{code}{an issue short code, e.g. ccm}
}
\description{
Lookup occurrence issue definitions and short codes
}
\examples{
occ_issues_lookup(issue = 'CONTINENT_COUNTRY_MISMATCH')
occ_issues_lookup(issue = 'MULTIMEDIA_DATE_INVALID')
occ_issues_lookup(issue = 'ZERO_COORDINATE')
occ_issues_lookup(code = 'cdiv')
}

