% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_term.r
\name{occ_term}
\alias{occ_term}
\title{Get occurrence terms.}
\usage{
occ_term(curlopts = list(http_version = 2))
}
\arguments{
\item{curlopts}{list of named curl options passed on to
\code{\link[crul]{HttpClient}}. see \code{curl::curl_options}
for curl options}
}
\value{
A \code{data.frame} with occurrence term definitions.
}
\description{
Get occurrence terms.
}
\examples{
\dontrun{
occ_term()

# Pass on curl options
occ_term(curlopts = list(verbose = TRUE))
}
}
\references{
\url{https://www.gbif.org/developer/occurrence}
}
