\name{occurrencedensity}
\alias{occurrencedensity}
\title{Returns summary counts of occurrence records by one-degree cell for a single
		taxon, country, dataset, data publisher or data network.}
\usage{
  occurrencedensity()
}
\description{
  Returns summary counts of occurrence records by
  one-degree cell for a single taxon, country, dataset,
  data publisher or data network.
}

