\name{occurrencelist_all}
\alias{occurrencelist_all}
\title{Occurrencelist_all carries out an occurrencelist query for a single name and all its name variants according to GBIF's name matching.}
\usage{
  occurrencelist_all()
}
\description{
  Note: This function is deprecated. See \code{occ_search}
}

