% Generated by roxygen2 (4.0.2): do not edit by hand
\name{organizations}
\alias{organizations}
\title{Organizations metadata.}
\usage{
organizations(data = "all", uuid = NULL, query = NULL, limit = 100,
  start = NULL, callopts = list())
}
\arguments{
\item{data}{The type of data to get. Default is all data.}

\item{uuid}{UUID of the data node provider. This must be specified if data
is anything other than 'all'.}

\item{query}{Query nodes. Only used when data='all'}

\item{limit}{Number of records to return. For \code{occ_search} this is 500, for others
its 100}

\item{start}{Record number to start at}

\item{callopts}{Pass on options to httr::GET for more refined control of
http calls, and error handling}
}
\value{
A list of length one or two. If \code{uuid} is NULL, then a data.frame with
call metadata, and a data.frame, but if \code{uuid} given, then a list.
}
\description{
Organizations metadata.
}
\examples{
\dontrun{
organizations(limit=5)
organizations(query="france")
organizations(uuid="4b4b2111-ee51-45f5-bf5e-f535f4a1c9dc")
organizations(data='contact', uuid="4b4b2111-ee51-45f5-bf5e-f535f4a1c9dc")
organizations(data='pending')
organizations(data=c('contact','endpoint'), uuid="4b4b2111-ee51-45f5-bf5e-f535f4a1c9dc")
}
}
\references{
\url{http://www.gbif.org/developer/registry#organizations}
}

