% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbif-package.r
\name{rgbif-defunct}
\alias{rgbif-defunct}
\title{Defunct functions in rgbif}
\description{
\itemize{
\item \code{\link[=density_spplist]{density_spplist()}}: service no longer provided
\item \code{\link[=densitylist]{densitylist()}}: service no longer provided
\item \code{\link[=gbifdata]{gbifdata()}}: service no longer provided
\item \code{\link[=gbifmap_dens]{gbifmap_dens()}}: service no longer provided
\item \code{\link[=gbifmap_list]{gbifmap_list()}}: service no longer provided
\item \code{\link[=occurrencedensity]{occurrencedensity()}}: service no longer provided
\item \code{\link[=providers]{providers()}}: service no longer provided
\item \code{\link[=resources]{resources()}}: service no longer provided
\item \code{\link[=taxoncount]{taxoncount()}}: service no longer provided
\item \code{\link[=taxonget]{taxonget()}}: service no longer provided
\item \code{\link[=taxonsearch]{taxonsearch()}}: service no longer provided
\item \code{\link[=stylegeojson]{stylegeojson()}}: moving this functionality to spocc package, will be
removed soon
\item \code{\link[=togeojson]{togeojson()}}: moving this functionality to spocc package, will be
removed soon
\item \code{\link[=gist]{gist()}}: moving this functionality to spocc package, will be
removed soon
\item \code{\link[=occ_spellcheck]{occ_spellcheck()}}: GBIF has removed the \code{spellCheck} parameter
from their API
}
}
\details{
The above functions have been removed. See
\url{https://github.com/ropensci/rgbif} and poke around the code if you
want to find the old functions in previous versions of the package
}
