% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{rgbif-package}
\alias{rgbif}
\alias{rgbif-package}
\title{Interface to the Global Biodiversity Information Facility API.}
\description{
rgbif: A programmatic interface to the Web Service methods provided by the
Global Biodiversity Information Facility.
}
\section{About}{


This package gives you access to data from GBIF \url{http://www.gbif.org/} via their API.
}

\section{A note about the old GBIF API}{


If you still want to use the old GBIF API, described at
\url{http://data.gbif.org/tutorial/services} let us know at
\url{https://github.com/ropensci/rgbif}, or email Scott \email{myrmecocystus@gmail.com}
}

\section{Documentation for the GBIF API}{


\itemize{
  \item summary \url{http://www.gbif.org/developer/summary}
  \item registry \url{http://www.gbif.org/developer/registry}
  \item species names \url{http://www.gbif.org/developer/species}
  \item occurrences \url{http://www.gbif.org/developer/occurrence}
  \item maps \url{http://www.gbif.org/developer/maps}
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}

Karthik Ram \email{karthik@ropensci.org}

Dan Mcglinn \email{danmcglinn@gmail.com}

Vijay Barve \email{vijay.barve@gmail.com}
}

