% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rgbif-package.r
\docType{package}
\name{rgbif-package}
\alias{rgbif}
\alias{rgbif-package}
\title{Interface to the Global Biodiversity Information Facility API.}
\description{
rgbif: A programmatic interface to the Web Service methods provided by the
Global Biodiversity Information Facility.
}
\section{About}{


This package gives you access to data from GBIF \url{http://www.gbif.org/} via their API.
}

\section{A note about the old GBIF API}{


The old GBIF API was at \url{http://data.gbif.org/tutorial/services}, but is now defunct -
that is, not available anymore. We used to have functions that worked with the old API, but
those functions are now not available anymore because GBIF made the old API defunct.
}

\section{Documentation for the GBIF API}{


\itemize{
  \item summary \url{http://www.gbif.org/developer/summary} - Summary of the GBIF API
  \item registry \url{http://www.gbif.org/developer/registry} - Metadata on datasets, and
  contributing organizations
  \item species names \url{http://www.gbif.org/developer/species} - Species names and metadata
  \item occurrences \url{http://www.gbif.org/developer/occurrence} - Occurrences
  \item maps \url{http://www.gbif.org/developer/maps} - Maps - these APIs are not implemented
  in \code{rgbif}, and are meant more for intergration with web based maps.
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}

Karthik Ram \email{karthik@ropensci.org}

Dan Mcglinn \email{danmcglinn@gmail.com}

Vijay Barve \email{vijay.barve@gmail.com}
}

