% Generated by roxygen2 (4.0.0): do not edit by hand
\name{taxrank}
\alias{taxrank}
\title{Get the possible values to be used for (taxonomic) rank arguments in GBIF
  	API methods.}
\usage{
taxrank()
}
\description{
Get the possible values to be used for (taxonomic) rank arguments in GBIF
  	API methods.
}
\examples{
\dontrun{
taxrank()
}
}

